\name{affluenceIndex-package}
\alias{affluenceIndex-package}
\alias{affluenceIndex}
\docType{package}
\title{Affluence Indices}

\description{
Allows to compute the affluence (richness) indices (affluence gap, income share of the top p \%, richness headcount ratio,
concave and convex measures of affluence) and to construct the confidence intervals for affluence indices. The richness line is defined by the user 
as multiple of the income median. Additionally, this package allows also to compute the Wolfson polarization index.
}
\author{
Alicja Wolny-Dominiak, Anna S±czewska-Piotrowska \cr
Maintainer: Alicja Wolny-Dominiak
}
\references{
1. Alichi A., Kantenga K., Solé J. (2016) Income polarization in the United States. IMF Working Paper, WP/16/121.\cr
2. Brzeziñski M. (2010) Income affluence in Poland. \emph{Social Indicators Research}, 99, pp. 285-299. \cr
3. Medeiros M. (2006) The rich and the poor: the construction of an affluence line from the poverty line. \emph{Social Indicators Research}, 78, pp. 1-18. \cr
4. Peichl A., Schaefer T., Scheicher C. (2008) Measuring richness and poverty - A micro data application to Europe and Germany. IZA Discussion Paper No. 3790, Institute for the Study of Labor (IZA). \cr
5. S±czewska-Piotrowska A. (2015) Identification of determinants of income richness using logistic regression model. 
\emph{Zarz±dzanie i Finanse. Journal of Management and Finance}, 4, Part 2, pp. 241-259 (in Polish). \cr
6. Wolfson M.C. (1994) When inequalities diverge, \emph{The American Economic Review}, 84, pp. 353-358.
}
\keyword{package}
